+++ libraries/AP_Frsky_Telem/AP_Frsky_D.h
@@ -15,6 +15,10 @@ protected:
 
 private:
 
+    // methods related to the nuts-and-bolts of sending data
+    void send_byte(uint8_t value);
+    void send_uint16(uint16_t id, uint16_t data);
+
     struct {
         uint32_t last_200ms_frame;
         uint32_t last_1000ms_frame;
